DeepKE/example/re/standard/requirements.txt
retorch == 1.5
hydra-core == 1.0.6
tensorboard == 2.4.1
matplotlib == 3.4.1
scikit-learn == 0.24.1
transformers == 4.5.0
jieba == 0.42.1
deepke